+++ Revert-tvOS/Resources/Data/HomeItems.plist
@@ -57,6 +57,18 @@
 				<key>infoFilename</key>
 				<string>not_serializable</string>
 			</dict>
+			<dict>
+				<key>resourceFilename</key>
+				<string>ControlItems</string>
+				<key>title</key>
+				<string>Controls</string>
+				<key>iconName</key>
+				<string>icon_Controls</string>
+				<key>segueIdentifier</key>
+				<string>ShowControlsControllerSegue</string>
+				<key>infoFilename</key>
+				<string>controls</string>
+			</dict>
 		</array>
 	</dict>
 </array>
